Design Tokens Manager - Refactor scale generator (#4510)
* refactoring and cleanups

* add unique key for NameValuePairs

* remove unused imports

* fix name value list
