+++ design-tokens-manager/src/routes/Scaler/components/ScaleLayout.jsx
@@ -1,0 +1,15 @@
+import { Box } from 'grommet';
+
+export const ScaleLayout = ({ ...rest }) => {
+  return (
+    <Box
+      direction="row"
+      align="end"
+      alignSelf="start"
+      cssGap
+      gap="medium"
+      wrap
+      {...rest}
+    />
+  );
+};
