+++ design-tokens-manager/src/routes/Scaler/components/ScaleValue.jsx
@@ -1,0 +1,21 @@
+import { Box, Text } from 'grommet';
+
+export const ScaleValue = ({ base, stops, value, ...rest }) => {
+  return (
+    <Box align="center" gap="xsmall" {...rest}>
+      <Box
+        background={
+          stops.includes(value)
+            ? { color: 'purple!', opacity: 'strong' }
+            : { color: 'purple!', opacity: 'medium' }
+        }
+        border={
+          value === base ? { color: 'orange!', size: 'small' } : undefined
+        }
+        width={`${value}px`}
+        height={`${value}px`}
+      />
+      <Text size="small">{value}px</Text>
+    </Box>
+  );
+};
