+++ design-tokens-manager/src/routes/Scaler/components/index.js
@@ -1,0 +1,5 @@
+export * from './CopyButton';
+export * from './ScaleLayout';
+export * from './ScaleToolbar';
+export * from './ScaleValue';
+export * from './SpacingValues';
